third_party/WebKit/LayoutTests/vr/exitPresent_reject_notpresenting.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="resources/fake-vr-displays.js"></script>
<script src="file:///gen/layout_test_data/mojo/public/js/mojo_bindings.js"></script>
<script src="file:///gen/device/vr/public/mojom/vr_service.mojom.js"></script>
<script src="resources/mock-vr-service.js"></script>
<canvas id="webgl-canvas"></canvas>
<script src="resources/presentation-setup.js"></script>
<script>
let fakeDisplays = fakeVRDisplays();

vr_test( (t) => {
  return navigator.getVRDisplays().then( (displays) => {
    var display = displays[0];

    return display.exitPresent().then( () => {
      t.step( () => {
        assert_unreached();
      }, "exitPresent succeeded unexpectedly");
    }, (err) => {
      t.step( () => {
        assert_true(err instanceof DOMException);
        assert_equals(err.name, "InvalidStateError");
        assert_false(display.isPresenting);
      }, "Request rejected properly");
    }).then( () => {
      t.done();
    });
  }, (err) => {
    t.step( () => {
      assert_unreached(err);
    }, "getVRDisplays rejected");
    t.done()
  });
}, [fakeDisplays["Pixel"]],
"Test exitPresent rejects if not presenting");

</script>
